del interface
