+++ src/main/java/com/github/t45k/abcd/output/OutputCSV.java
@@ -2,7 +2,7 @@ package com.github.t45k.abcd.output;
 
 import com.github.t45k.abcd.clone.entity.CloneSet;
 
-public class OutputCSV extends AbstractOutput {
+public class OutputCSV extends Output {
 
     @Override
     protected String getExtension() {
