+++ src/main/java/com/github/t45k/abcd/output/OutputJson.java
@@ -2,7 +2,7 @@ package com.github.t45k.abcd.output;
 
 import com.github.t45k.abcd.clone.entity.CloneSet;
 
-public class OutputJson extends AbstractOutput {
+public class OutputJson extends Output {
 
     @Override
     protected String getExtension() {
