+++ src/main/java/com/github/t45k/abcd/output/OutputXML.java
@@ -2,7 +2,7 @@ package com.github.t45k.abcd.output;
 
 import com.github.t45k.abcd.clone.entity.CloneSet;
 
-public class OutputXML extends AbstractOutput {
+public class OutputXML extends Output {
 
     @Override
     protected String getExtension() {
